Move DeltaTime and WindowSize resources to separate files
