+++ src/resources/window_size.rs
@@ -1,0 +1,2 @@
+#[derive(Default)]
+pub struct WindowSize(pub (u32, u32));
